feat: pull out otel helpers and use them to test the many example
