+++ rust/tests/many_tests.rs
@@ -1,9 +1,14 @@
 #[cfg(test)]
 mod tests {
     use anyhow::Result;
+    use serde_json::Value;
+    use std::convert::identity;
     use std::process::Command;
     use std::thread;
     use std::time;
+
+    mod helpers;
+    use helpers::otel_json::*;
 
     #[test]
     fn integration_many() -> Result<()> {
@@ -15,8 +20,8 @@ mod tests {
         thread::sleep(time::Duration::from_millis(150));
         let output = String::from_utf8(output.stdout)?;
         let output_lines = output.lines();
-        println!("{}", output);
-        // First test that the expected output was printed 10 times
+
+        // First test that the modules ran the expected number of times
         assert_eq!(
             output_lines
                 .clone()
@@ -25,15 +30,20 @@ mod tests {
             1000
         );
 
-        // Check that printf was also called 10 times
-        assert_eq!(
-            output_lines
-                .clone()
-                .filter(|l| l.contains("printf"))
-                .count(),
-            100
-        );
-
+        // check that every allocation was called
+        let traces = output_lines
+            .map(|l| match serde_json::from_str(l) {
+                Ok(x) => Some(x),
+                Err(_) => None,
+            })
+            .collect::<Vec<Option<Value>>>()
+            .into_iter()
+            .filter_map(identity)
+            .collect::<Vec<Value>>();
+        let allocations = traces
+            .iter()
+            .filter(|t| attribute_of_first_span(t, "name".to_string()).unwrap() == "allocation");
+        assert_eq!(allocations.count(), 1000);
         Ok(())
     }
 }
