+++ rust/tests/tests/helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod otel_json;
